requirements.txt
asgiref==3.9.1
dal-admin-filters==1.1.0
Django==5.2.5
django-autocomplete-light==3.12.1
django-debug-toolbar==6.0.0
django-filter==25.1
django-polymorphic==4.1.0
django-taggit==6.1.0
Faker==37.6.0
model-bakery==1.20.5
sqlparse==0.5.3
tzdata==2025.2
